# Make getDimensionFromEmptyDims allocate less memory