Make getDimensionFromName allocate less memory:
```
mwp@Matthews-MacBook-Pro-3 ~/.go/src/github.com/signalfx/metricproxy (enhance)*$ benchcmp /tmp/old.txt  /tmp/newer.txt
benchmark                                    old ns/op     new ns/op     delta
BenchmarkGetDimensionFromNameEmptyDims-8     1826          1532          -16.10%
BenchmarkGetDimensionFromName-8              3679          3184          -13.45%

benchmark                                    old allocs     new allocs     delta
BenchmarkGetDimensionFromNameEmptyDims-8     4              2              -50.00%
BenchmarkGetDimensionFromName-8              9              6              -33.33%

benchmark                                    old bytes     new bytes     delta
BenchmarkGetDimensionFromNameEmptyDims-8     159           63            -60.38%
BenchmarkGetDimensionFromName-8              575           463           -19.48%
```

This is used for every plugin_instance, type_instance or host dimension passed in from collectd